Shooting, colored enemies, flashy gems and some bug fixes.
Implemented the shooting mechanism. It doesn't wrap but what does!!
A couple more gems have been added with the ability to flash.
All enemies now start with a random ZX Spectrum color.
Extra decrementation of the InBubble counter was happening when I added extra bullets (as the following bullet from the one that killed the enemy also registered as a kill!!). This caused the InBubble counter to go into th -ve and now allow respawning of the player. This is now fixed.