Game over UI, buttons and level initialisation fixes.
Implemented the Game Over screen and fixed a few level restart variable initialisation errors.
Moved the PlayerSpawnPoint into the GameMaster object as it was losing the reference to it in the GameMaster script between levels. Don't know why but this has fixed it.
Removed the PlayerSpawnPoint reference in the LevelController script.
Added a ButtonController to control what the buttons do!!