Repawn sound and refactor
Implemented the respawn sound.
Refactored the SpawnSpaceman() methods so a delay value is passed as a parameter rather than hardcoded into the method.
Removed the spaceman from the original scene and have him "respawn" at the start of the game to get the sound.
Added a respawn into the InitialiseScoresAndLives() method to make sure the player is respawned when starting a new sub-level.