Explosion to separate object & respawn fix.
Moved the explosion when enemies doe to a separate object. Now the player can fly through the explosion without being killed.
EnemiesInBubble was being doubly decremented as the ExitTrigger decrementing was happening when they were killed along with the Bullet/Player decrementing. It only gets decremented in the ExitTrigger method now.